refactor(main): remove git protocol support
As this is still used by the test logic, I've moved it to the _test.go
files. I don't expect to need this in production as the plaintext git
protocol is rather insecure.
